+++ src/portable/extension.rs
@@ -3,20 +3,18 @@ use std::path::Path;
 use std::process::Command;
 
 use anyhow::Context;
+use edgedb_cli_derive::IntoArgs;
 use log::trace;
 use prettytable::{row, Table};
 
 use crate::branding::BRANDING_CLOUD;
 use crate::hint::HintExt;
 use crate::options::Options;
-use crate::portable::install::download_package;
 use crate::portable::local::InstanceInfo;
-use crate::portable::options::{
-    instance_arg, ExtensionCommand, ExtensionInstall, ExtensionList, ExtensionListAvailable,
-    ExtensionUninstall, InstanceExtensionCommand, InstanceName,
-};
+use crate::portable::options::{instance_arg, InstanceName};
 use crate::portable::platform::get_server;
 use crate::portable::repository::{get_platform_extension_packages, Channel};
+use crate::portable::server::install::download_package;
 use crate::table;
 
 pub fn run(c: &ExtensionCommand, o: &Options) -> Result<(), anyhow::Error> {
@@ -27,6 +25,79 @@ pub fn run(c: &ExtensionCommand, o: &Options) -> Result<(), anyhow::Error> {
         ListAvailable(c) => list_available(c, o),
         Uninstall(c) => uninstall(c, o),
     }
+}
+
+#[derive(clap::Args, Debug, Clone)]
+#[command(version = "help_expand")]
+#[command(disable_version_flag = true)]
+pub struct ExtensionCommand {
+    #[command(subcommand)]
+    pub subcommand: InstanceExtensionCommand,
+
+    /// Name of the instance
+    #[arg(short = 'I', long)]
+    #[arg(global = true)]
+    pub instance: Option<InstanceName>,
+}
+
+#[derive(clap::Subcommand, Clone, Debug)]
+pub enum InstanceExtensionCommand {
+    /// List installed extensions for a local instance.
+    List(ExtensionList),
+    /// List available extensions for a local instance.
+    ListAvailable(ExtensionListAvailable),
+    /// Install an extension for a local instance.
+    Install(ExtensionInstall),
+    /// Uninstall an extension from a local instance.
+    Uninstall(ExtensionUninstall),
+}
+
+#[derive(clap::Args, Debug, Clone)]
+pub struct ExtensionList {
+    #[arg(from_global)]
+    pub instance: Option<InstanceName>,
+}
+
+#[derive(clap::Args, IntoArgs, Debug, Clone)]
+pub struct ExtensionListAvailable {
+    /// Specify the channel override (stable, testing, or nightly)
+    #[arg(long, hide = true)]
+    pub channel: Option<Channel>,
+    /// Specify the slot override (for development use)
+    #[arg(long, hide = true)]
+    pub slot: Option<String>,
+
+    #[arg(from_global)]
+    pub instance: Option<InstanceName>,
+}
+
+#[derive(clap::Args, IntoArgs, Debug, Clone)]
+pub struct ExtensionInstall {
+    #[arg(from_global)]
+    pub instance: Option<InstanceName>,
+
+    /// Name of the extension to install
+    #[arg(short = 'E', long)]
+    pub extension: String,
+    /// Specify the channel override (stable, testing, or nightly)
+    #[arg(long, hide = true)]
+    pub channel: Option<Channel>,
+    /// Specify the slot override (for development use)
+    #[arg(long, hide = true)]
+    pub slot: Option<String>,
+    /// Reinstall the extension if it's already installed
+    #[arg(long, hide = true)]
+    pub reinstall: bool,
+}
+/// Represents the options for uninstalling an extension from a local EdgeDB instance.
+#[derive(clap::Args, IntoArgs, Debug, Clone)]
+pub struct ExtensionUninstall {
+    #[arg(from_global)]
+    pub instance: Option<InstanceName>,
+
+    /// The name of the extension to uninstall.
+    #[arg(short = 'E', long)]
+    pub extension: String,
 }
 
 fn get_local_instance(instance: &Option<InstanceName>) -> Result<InstanceInfo, anyhow::Error> {
