+++ src/portable/main.rs
@@ -1,5 +1,5 @@
 use crate::options::Options;
-use crate::portable::options::{ServerCommand, ServerInstanceCommand};
+use crate::portable::options::ServerInstanceCommand;
 use crate::portable::project::ProjectCommand;
 
 use crate::portable::backup;
@@ -7,33 +7,14 @@ use crate::portable::control;
 use crate::portable::create;
 use crate::portable::credentials;
 use crate::portable::destroy;
-use crate::portable::info;
-use crate::portable::install;
 use crate::portable::link;
-use crate::portable::list_versions;
 use crate::portable::project;
 use crate::portable::reset_password;
 use crate::portable::resize;
 use crate::portable::revert;
 use crate::portable::status;
-use crate::portable::uninstall;
 use crate::portable::upgrade;
 use crate::portable::windows;
-
-pub fn server_main(cmd: &ServerCommand) -> Result<(), anyhow::Error> {
-    use crate::portable::options::Command::*;
-
-    match &cmd.subcommand {
-        Install(c) if cfg!(windows) => windows::install(c),
-        Install(c) => install::install(c),
-        Uninstall(c) if cfg!(windows) => windows::uninstall(c),
-        Uninstall(c) => uninstall::uninstall(c),
-        ListVersions(c) if cfg!(windows) => windows::list_versions(c),
-        ListVersions(c) => list_versions::list_versions(c),
-        Info(c) if cfg!(windows) => windows::info(c),
-        Info(c) => info::info(c),
-    }
-}
 
 pub fn instance_main(cmd: &ServerInstanceCommand, options: &Options) -> Result<(), anyhow::Error> {
     use crate::portable::options::InstanceCommand::*;
