+++ src/portable/mod.rs
@@ -17,17 +17,14 @@ mod create;
 mod credentials;
 mod destroy;
 pub mod extension;
-mod info;
-pub mod install;
 mod link;
-mod list_versions;
 pub mod project;
 mod reset_password;
 mod resize;
 mod revert;
+pub mod server;
 pub mod status;
-mod uninstall;
 mod upgrade;
 
-pub use main::{instance_main, project_main, server_main};
+pub use main::{instance_main, project_main};
 pub use reset_password::password_hash;
