+++ src/portable/options.rs
@@ -23,12 +23,6 @@ const DOMAIN_LABEL_MAX_LENGTH: usize = 63;
 const CLOUD_INSTANCE_NAME_MAX_LENGTH: usize = DOMAIN_LABEL_MAX_LENGTH - 2 + 1; // "--" -> "/"
 
 #[derive(clap::Args, Debug, Clone)]
-pub struct ServerCommand {
-    #[command(subcommand)]
-    pub subcommand: Command,
-}
-
-#[derive(clap::Args, Debug, Clone)]
 #[command(version = "help_expand")]
 #[command(disable_version_flag = true)]
 pub struct ServerInstanceCommand {
@@ -83,176 +77,6 @@ pub enum InstanceCommand {
     ResetPassword(ResetPassword),
     /// Display instance credentials (add `--json` for verbose).
     Credentials(ShowCredentials),
-}
-
-#[derive(clap::Args, Debug, Clone)]
-#[command(version = "help_expand")]
-#[command(disable_version_flag = true)]
-pub struct ExtensionCommand {
-    #[command(subcommand)]
-    pub subcommand: InstanceExtensionCommand,
-
-    /// Name of the instance
-    #[arg(short = 'I', long)]
-    #[arg(global = true)]
-    pub instance: Option<InstanceName>,
-}
-
-#[derive(clap::Subcommand, Clone, Debug)]
-pub enum InstanceExtensionCommand {
-    /// List installed extensions for a local instance.
-    List(ExtensionList),
-    /// List available extensions for a local instance.
-    ListAvailable(ExtensionListAvailable),
-    /// Install an extension for a local instance.
-    Install(ExtensionInstall),
-    /// Uninstall an extension from a local instance.
-    Uninstall(ExtensionUninstall),
-}
-
-#[derive(clap::Args, Debug, Clone)]
-pub struct ExtensionList {
-    #[arg(from_global)]
-    pub instance: Option<InstanceName>,
-}
-
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct ExtensionListAvailable {
-    /// Specify the channel override (stable, testing, or nightly)
-    #[arg(long, hide = true)]
-    pub channel: Option<Channel>,
-    /// Specify the slot override (for development use)
-    #[arg(long, hide = true)]
-    pub slot: Option<String>,
-
-    #[arg(from_global)]
-    pub instance: Option<InstanceName>,
-}
-
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct ExtensionInstall {
-    #[arg(from_global)]
-    pub instance: Option<InstanceName>,
-
-    /// Name of the extension to install
-    #[arg(short = 'E', long)]
-    pub extension: String,
-    /// Specify the channel override (stable, testing, or nightly)
-    #[arg(long, hide = true)]
-    pub channel: Option<Channel>,
-    /// Specify the slot override (for development use)
-    #[arg(long, hide = true)]
-    pub slot: Option<String>,
-    /// Reinstall the extension if it's already installed
-    #[arg(long, hide = true)]
-    pub reinstall: bool,
-}
-/// Represents the options for uninstalling an extension from a local EdgeDB instance.
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct ExtensionUninstall {
-    #[arg(from_global)]
-    pub instance: Option<InstanceName>,
-
-    /// The name of the extension to uninstall.
-    #[arg(short = 'E', long)]
-    pub extension: String,
-}
-
-#[derive(clap::Subcommand, Clone, Debug)]
-pub enum Command {
-    /// Show locally installed server versions.
-    Info(Info),
-    /// Install a server version locally.
-    Install(Install),
-    /// Uninstall a server version locally.
-    Uninstall(Uninstall),
-    /// List available and installed versions of the server.
-    ListVersions(ListVersions),
-}
-
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct Info {
-    /// Display only the server binary path (shortcut to `--get bin-path`).
-    #[arg(long)]
-    pub bin_path: bool,
-    /// Output in JSON format.
-    #[arg(long)]
-    pub json: bool,
-
-    // Display info for latest version.
-    #[arg(long)]
-    #[arg(conflicts_with_all=&["channel", "version", "nightly"])]
-    pub latest: bool,
-    // Display info for nightly version.
-    #[arg(long)]
-    #[arg(conflicts_with_all=&["channel", "version", "latest"])]
-    pub nightly: bool,
-    // Display info for specific version.
-    #[arg(long)]
-    #[arg(conflicts_with_all=&["nightly", "channel", "latest"])]
-    pub version: Option<ver::Filter>,
-    // Display info for specific channel.
-    #[arg(long, value_enum)]
-    #[arg(conflicts_with_all=&["nightly", "version", "latest"])]
-    pub channel: Option<Channel>,
-
-    /// Get specific value:
-    ///
-    /// * `bin-path` -- Path to the server binary
-    /// * `version` -- Server version
-    #[arg(long, value_parser=["bin-path", "version"])]
-    pub get: Option<String>,
-}
-
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct Install {
-    #[arg(short = 'i', long)]
-    pub interactive: bool,
-    #[arg(long, conflicts_with_all=&["channel", "version"])]
-    pub nightly: bool,
-    #[arg(long, conflicts_with_all=&["nightly", "channel"])]
-    pub version: Option<ver::Filter>,
-    #[arg(long, conflicts_with_all=&["nightly", "version"])]
-    #[arg(value_enum)]
-    pub channel: Option<Channel>,
-}
-
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct Uninstall {
-    /// Uninstall all versions.
-    #[arg(long)]
-    pub all: bool,
-    /// Uninstall unused versions.
-    #[arg(long)]
-    pub unused: bool,
-    /// Uninstall nightly versions.
-    #[arg(long, conflicts_with_all=&["channel"])]
-    pub nightly: bool,
-    /// Uninstall specific version.
-    pub version: Option<String>,
-    /// Uninstall only versions from a specific channel.
-    #[arg(long, conflicts_with_all=&["nightly"])]
-    #[arg(value_enum)]
-    pub channel: Option<Channel>,
-    /// Increase verbosity.
-    #[arg(short = 'v', long)]
-    pub verbose: bool,
-}
-
-#[derive(clap::Args, IntoArgs, Debug, Clone)]
-pub struct ListVersions {
-    #[arg(long)]
-    pub installed_only: bool,
-
-    /// Single column output.
-    #[arg(long, value_parser=[
-        "major-version", "installed", "available",
-    ])]
-    pub column: Option<String>,
-
-    /// Output in JSON format.
-    #[arg(long)]
-    pub json: bool,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize, clap::ValueEnum)]
