+++ src/portable/server/uninstall.rs
@@ -6,8 +6,8 @@ use crate::platform::{data_dir, portable_dir, tmp_file_path};
 use crate::portable::exit_codes;
 use crate::portable::local;
 use crate::portable::local::InstanceInfo;
-use crate::portable::options::Uninstall;
 use crate::portable::repository::Query;
+use crate::portable::server::Uninstall;
 use crate::portable::status;
 use crate::portable::ver;
 use crate::print::{self, msg, Highlight};
