+++ src/portable/windows.rs
@@ -31,6 +31,7 @@ use crate::portable::local::{write_json, InstanceInfo, NonLocalInstance, Paths};
 use crate::portable::options::{self, instance_arg, InstanceName, Logs, StartConf};
 use crate::portable::project;
 use crate::portable::repository::{self, download, PackageHash, PackageInfo};
+use crate::portable::server;
 use crate::portable::status::{self, Service};
 use crate::portable::ver;
 use crate::print::{self, msg, Highlight};
@@ -711,7 +712,7 @@ pub fn is_wrapped() -> bool {
     Env::_from_windows().unwrap_or_default().unwrap_or_default()
 }
 
-pub fn install(options: &options::Install) -> anyhow::Result<()> {
+pub fn install(options: &server::Install) -> anyhow::Result<()> {
     ensure_wsl()?
         .edgedb()
         .arg("server")
@@ -721,7 +722,7 @@ pub fn install(options: &options::Install) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn uninstall(options: &options::Uninstall) -> anyhow::Result<()> {
+pub fn uninstall(options: &server::Uninstall) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         wsl.edgedb()
             .arg("server")
@@ -737,7 +738,7 @@ pub fn uninstall(options: &options::Uninstall) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn list_versions(options: &options::ListVersions) -> anyhow::Result<()> {
+pub fn list_versions(options: &server::ListVersions) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         wsl.edgedb()
             .arg("server")
@@ -755,7 +756,7 @@ pub fn list_versions(options: &options::ListVersions) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn info(options: &options::Info) -> anyhow::Result<()> {
+pub fn info(options: &server::Info) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         wsl.edgedb().arg("server").arg("info").args(options).run()?;
     } else {
